Update patch set 1
Patch Set 1:

(3 comments)

This is a cleanup patch from a series of audited tests that was deferred until all other features were done.

The dmesg -T changes to use Android Time was to allow for an easier check of that, but it was rejected. Until we have the ability to take 'dmesg -T' and compare 1:1 against 'logcat -b kernel -v usec -d' automated unit tests will not occur. A wrapper for dmesg -T to perhaps simulate that has yet to be written; in time ...

And in the purest send, can not write an automated unit test without adding code to allow test-input (replace klogctl), the input comes from the kernel buffer. My unit test has been to allow a boot and multiple devices over time, then to confirm that all log lines are in their place, tested on a variety of devices to cover all the oddities that could come from the kernel. The audits so far are done visually because kernel logs do not shape well into the android logs (the content is not 1:1 exact match because of spacing rules).

Patch-set: 1