Update patch set 2
Patch Set 2:

(1 comment)

> Seems fine, but are you sure you want to encode this functionality
 > in init.rc instead of in init?  Why not have keywords to specify
 > the behavior that you want, and then have init take care of the
 > details?  That way you can change the implementation later without
 > having to find every device that uses it an update their init.rc.

I think more complicated cgroup policy should probably be considered device-specific for the foreseeable future. this also doesn't require us to encode a lot of policy into init. I don't think it's a burden for now, but I agree that if some basic patterns for services emerge we may want to include support for that in init directly.

Patch-set: 2