Update patch set 2
Patch Set 2:

Is this change really dependent on https://android-review.googlesource.com/161114? I tested this change against the old AdbWinUsbApi.dll (without 161114) and in practice it actually seems to fix the hang problem, however, there is still a chance of hitting a crash race condition without 161114.

In other words, if someone somehow mismatches adb.exe and AdbWinUsbApi.dll, it is suboptimal, but not 100% immediately disastrous. It's still best to use both fixes (161114 and this).

Patch-set: 2