Update patch set 3
Patch Set 3:

> i still haven't understood your comment here --- are you saying
 > that without the device driver fix, this patch can actually make
 > things worse? 

Well, when I tested it, this patch seemed to make things better even without the device driver fix. The caveat is that this patch introduces a new thread (_power_notification_thread) that calls AdbCloseHandle() on resume. But that call to AdbCloseHandle() could race with the other threads. With the device driver fix, the race doesn't matter because it fixes all races, but without that fix, this patch adds another chance for the race to crash.

But the alternative to not having this patch is adb hanging on resume. IMO, it's better to take this patch, risk the chance of crash instead of being like me a few months ago constantly replugging cables. :) But the call is yours to make.

 > or just that without the device driver fix, this
 > patch doesn't solve everything?

Yes, exactly, this patch doesn't solve everything. To get 'everything' you also need the device driver fix.

 > (because another week's gone by without me finding time to look at
 > the DDK, so if this offers some benefit and no downside without the
 > other change, i'd like to get it in.)

I think the conclusion is that there is some potential downside, but the upside is definitely very big.

If you need more time on these patches, maybe you should take it. These changes are kind of scary to begin with.

BTW, the other change is currently built with the DDK, but since the device driver source code has been deleted, AdbWinUsbApi.dll and AdbWinApi.dll no longer inherently need the DDK to be built. You could in theory get it building with the regular toolset. You might just need small replacements for some code that uses the ATL from Microsoft (the ATL is included with the DDK, but probably not part of mingw), but that is probably doable.

Patch-set: 3