Update patch set 3
Patch Set 3:

I dont agree fully. With domain transitions off of labeled filesystems  they can incorrectly run in the init domain now.

The problem isn't with the neverallows. Tthe problem is with the vast teams porting this with varried experience and the inherent race condition they encounter with services coming up before policy porting and the seclabel's being ported ahead of time versus the policy. The day seclabel dies, and stops scattering the policy via hardcoded types, the easier porting will be and this change wouldn't be neccesary.

Patch-set: 3