Update patch set 3
Patch Set 3:

I see that you check the computation of the child process type in the patch shown and abort service launch if its the same.

I think theirs the theory and then the practice, where the porting teams in most of these OEMs are going to keep falling flat on their face with this for at least 2 more Android releases.

Patch-set: 3