Update patch set 2
Patch Set 2:

PS2 fixes kill-server to return 0 if the server isn't around (this is the old behavior that makes sense). If kill-server has some other error, the exit code is 1 (like before).

Thus, beyond having improved error output, the big change here is to output "* server not running *" if the server wasn't running to begin with. Previously, adb would output that when there was an error (??).

Patch-set: 2