Update patch set 2
Patch Set 2:

I believe this is mostly (but potentially not entirely) taken care of in the RestartIfNeeded() function.  Start() is only called for a process needing restart if (time_started_ + 5 <= gettime()).  I tested this with both adbd and mediaserver and see that when I kill them through adb, they wait 5 seconds since their last launch before they start.

Surfaceflinger and a few other services have onrestart commands, which explicitly restart a set of services, but Restart() also respects that if SVC_RESTARTING is already set, which happens in the signal handler (Reap()), that it should not Start() the service and should instead let RestartIfNeeded() work, which includes the delay.

Are you seeing a different route for services to be explicitly killed and restarted?  Maybe via the control mechanism (issuing start/stop on the console) repeated?

If so, I think we should adapt the mechanism used in RestartIfNeeded(), either having the next_start_time that RestartIfNeeded() uses as a member variable, or adapting it to use your restart_delay_ variable.

Patch-set: 2
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0