Update patch set 2
Patch Set 2: Code-Review+1

(2 comments)

Thanks for responding to my comments (some of them incorrect :)). I looked at my old notes on this bug and it seems like the 'real' issue was the call to transport_unref that was releasing a refcount that it didn't have, and that was causing all the problems. The other stuff was just a red herring.

Thanks for fixing my fix.

Patch-set: 2
Label: Code-Review=+1