Update patch set 3
Abandoned

Yabin fixed the crash/memory-corruption issue in https://android-review.googlesource.com/168492.

The BUGBUG on acquire_one_transport is not a bug because the transport stuff happens on one thread.

The Google guys can open bugs on the BUGBUGs in register_socket_transport and unregister_usb_transport if they are deemed to be true.

Patch-set: 3
Status: abandoned