Update patch set 3
Patch Set 3:

To clarify, I don't have the envionment to test on latest adb. Only reproduced and tested on LP5.1.

What I observed on LP5.1 is that on write error due to adb disconnection, no SIGPIPE handled, only -EPIPE returned on write(), but the problem in previous logic is that then it directly jump to wait for screencap process. But on this time screencap process still has not finished data output. Thus waitpid() can never return.

Patch-set: 3