Update patch set 3
Patch Set 3:

(2 comments)

> A few small comments, but as an overall suggestion I believe we do
 > have those Chromium conversion functions available in AOSP at
 > external/libchrome/base/strings/ - would it make more sense to
 > leave base/ alone here and instead just link to those existing
 > functions in adb?

It seems like there are some alternatives:

0. adb is calling android::base::ReadFileToString() which is the problem, so linking to libchrome's UTF8ToWide() doesn't seem like it would solve the direct problem.
1. Libchrome does have a base::ReadFileToString() method which could be called instead, though calling it would require using libchrome's FilePath class.
2. Libbase for Windows could take a dependency on libchrome and then android::base::ReadFileToString() could be fixed by having u8::open() use base::UTF8ToWide() from libchrome. Perhaps this is what you are suggesting?

I'm not opposed to using libchrome (especially since it has a lot of useful stuff across the board that I'd be happy to call), but pulling it in seems a bit more complicated. I'm not even sure why libchrome is there and for what primary purpose, so I don't really know if it makes sense to take a dependency on it. Perhaps linking to libchrome is something that a Googler should do instead?

Patch-set: 3
Label: Verified=0