Update patch set 4
Patch Set 4:

(1 comment)

Thanks a ton for the info and that reference, it was super helpful!

Re: performance, my impression of slowness was from a comment in Chromium's utf_string_conversions.h:16:

  // These convert between UTF-8, -16, and -32 strings. They are potentially slow,
  // so avoid unnecessary conversions.

but that comment is more general than our case here so it may not apply, I'd guess that you're right and we don't need to worry about performance in this case.

Patch-set: 4