Update patch set 4
Patch Set 4: Code-Review+1

OK thanks for trying it out anyway, libchrome is pretty new to AOSP and I had hoped that we had gotten it working on the host but it must only have been tested on the device side so far.

In that case this seems to me like the best solution to get Windows working again.

Patch-set: 4
Label: Code-Review=+1