Update patch set 1
Patch Set 1:

(1 comment)

Note that if you resize the console window when mksh is running, you get some weird text garbage:

https://drive.google.com/open?id=0B3A4jzWKJNSFbTBlcno4MTdlRnc

This also occurs on Ubuntu, but to a lesser extent:

https://drive.google.com/open?id=0B3A4jzWKJNSFVy10TVIzV0ZqMGc

What is going on is that when mksh gets SIGWINCH, it outputs "\r<prompt>               \b\b\b\b\b\b\b\b" with enough spaces and \b to match the number of columns. The problem is that it can send that output to the console window, which can already be resized to be smaller. So the text that mksh sent doesn't fit and the console has to wrap it to fit. Apparently Ubuntu wraps the content to another line, but Windows does something that looks a bit stranger.

I don't see an easy fix to this, but it just seems like something to live with. I'm open to ideas.

Patch-set: 1
Label: Verified=0