Update patch set 2
Patch Set 2:

I'm not totally opposed to making _console_read() maintain an input queue buffer since it already has the global _win32_input_record.

That being said, how come the memmove() idea doesn't work? Yes, there can be more than one ~ in the buffer, but if it is in the buffer, you memmove() on top of it, freeing space, then if the following char is not '.', you memmove() forward, consuming the space that was freed. So multiple ~ does not need more space, unless the buffer ends with ~, then the next read() will need to insert 1 ~ in the worst case.

Or am I misunderstanding?

Patch-set: 2