Update patch set 15
Patch Set 15: Code-Review+1 Verified+1

Rebase with a small set of scale improvements. Still do not have enough bravery to consider releasing this. Stable, Scientifically sound, appears to work as designed, but have not confirmed that the savings dealing with splitting out the sources of logspam in the system UID are dramatic enough for the added complexity.

Patch-set: 15
Label: Code-Review=+1
Label: Verified=+1