Update patch set 1
Patch Set 1:

> Would you happen to have any preference to how to move
 > ParseCompleteUTF8<>() around so that it can be called by a adb_test
 > sysdeps_win32_test.cpp unittest? Put the whole thing in sysdeps.h
 > (might be hard with dependencies)? Put two wrappers
 > ParseCompleteUTF8Array() and ParseCompleteUTF8Vector() in
 > sysdeps_win32.cpp and put their prototypes in sysdeps.h and call
 > them from the unittest?


Oh right templates make this a bit tricky. Adding a wrapper for the test seems reasonable to me, even just one of them would be fine, I don't think we need to test both types. Maybe just put the prototype in "namespace internal" or something so it's obvious that it shouldn't be called from anywhere else.

Alternately you could remove the templating and always take char* first/last or char* first and a size as inputs. With just 2 call sites the templating doesn't seem to buy us a whole lot here, but it's kind of cool and does make it a bit more flexible.

Patch-set: 1