Update patch set 1
Patch Set 1:

(2 comments)

> So if sending gets any sort of error, you'll do a blocking read? Is there any chance the daemon is still waiting and will never send anything, hanging the client forever?

i certainly think that's a possibility, but you can always hit ^C. i think for this to happen you'd need the write to fail locally (on the host) in a way that didn't imply any problem on the remote side. i'm struggling to imagine what that would be. if we find that's a real possibility, i think we can just shutdown(2) on our end before trying the read.

Patch-set: 1