Update patch set 1
Patch Set 1:

> I don't think I follow this. I thought SO_LINGER is used to cause a
 > RST packet to be sent, which causes ECONNRESET on the other end.
 > But the other end of the sync service socket is really in
 > sockets.cpp, so sockets.cpp will get the ECONNRESET, not the
 > client. In other words, I don't think all the socket code in adb
 > 'propagate' the ECONNRESET. (I think the chain is really: sync
 > service socket -> sockets.cpp -> transport thread -> USB -> adb
 > server transport thread -> sockets.cpp -> adb client)

no, i'm actually just using SO_LINGER for the opposite side-effect. _without_ it, Linux will send a RST if you keep writing to the closed socket. that's our current problem. we can't wait until we get ECONNRESET and then read adbd's error because of that RST. (i don't actually know whether it's because it's still buffered remotely and the device kernel doesn't send it, or the local buffer gets discarded.)

if i use SO_LINGER, i can say "give me 20s to get round to looking at the failure". so, yeah, i'm assuming that i'll stop writing in 20s, but at least for local devices and adb-sync-sized files, that's fine. and certainly better than right now, where we don't tell you what made the adbd unhappy, and just report that we're no longer talking to it.

 > Even if you can get the ECONNRESET propagation to work, I'm really
 > not sure if this is the best way to do things. Can't the socket
 > client & server just always do graceful shutdown? In other words,
 > after finishing writing, even if you get an error, do
 > shutdown(SHUT_WR) (which will cause the other side to break out of
 > any potential read()), then read() until 0? Will that be reliable
 > and prevent the hangs that we were worried about?

no. i tried that first. (remember i was asking about shutdown on Windows?)

the only other thing i think would work (but haven't tried) is poll(2)ing in the client to see if there's anything to read before i write.

 > Maybe reading the HTTP PUT spec will give some ideas?

i thought that just required that you accept the whole input? which would be an option, but only fix error reporting for > M devices. (and still has the same flaw as the current patch, which is that you have to send 20MiB of icu.dat before you can find out that the system partition is read-only.)

Patch-set: 1