Update patch set 3
Patch Set 3:

> The S_IS* macros being different between adbd and the adb client
 > seems pretty scary. We use a bunch of the other macros (S_ISREG,
 > S_ISDIR, etc.) in the client code against modes that get sent over
 > from adbd, so if we add S_IS_REMOTE_LNK, we should probably have
 > REMOTE versions of all of the macros. It looks like the Windows
 > definitions never conflict with the Linux ones, though, so maybe we
 > can just define S_ISLINK and S_IFLNK to match the Linux
 > definitions? Presumably (although this probably needs to be
 > verified), those bits will always be clear on Windows, so the
 > behavior shouldn't change.

Yeah, I think keeping S_ISLNK, etc. the same on Windows is another good way to solve this. It's fine with me. The only weird thing is what should happen if S_ISLNK is used on a local file on Windows, but I think that is already covered via #ifdef _WIN32 and so forth.

Patch-set: 3