Update patch set 1
Patch Set 1:

> The test test_non_interactive_sigint seems to fail if adb is not
 > running as root:
 > 
 > Traceback (most recent call last):
 > File "test_device.py", line 450, in test_non_interactive_sigint
 > 'subprocess failed to terminate')
 > AssertionError: subprocess failed to terminate
 > 
 > Is this expected? I don't recall hitting this issue a few months
 > ago, but maybe I'm mistaken.
 > 
 > My repro steps:
 > 
 > 1. Build adbd.
 > 2. Run API 23 emulator.
 > 3. Copy newly built adbd to /sbin/adbd
 > 4. Restart adbd with "setprop ctl.restart adbd"
 > 5. Use adb root or adb unroot to switch between root/shell.
 > 6. python -m unittest -v test_device.ShellTest.test_non_interactive_sigint

I've been unable to reproduce this on a physical device or emulator, though I have not been exactly following your steps as I am not sure how to copy to /sbin/adbd. But with a fresh build this seems to work fine.

A hunch is that if you're just updating adbd but not the rest of the system, adbd doesn't have the right permissions to kill the process (see http://r.android.com/171014). In adb/daemon/main.cpp it looks like permissions are only dropped for non-root builds, which would make root work even without the updated permissions file.

Patch-set: 1