Update patch set 4
Patch Set 4:

(1 comment)

As for landing it in system/core, we actually just decided to have it in a separate repo (just like update_engine) so e.g. bvbtool can easily be used outside the Android tree. I'll file bugs to create this repo and will port this and other CLs to it. I'll follow up when that's done.

Patch-set: 4