Update patch set 1
Patch Set 1:

> Windows stuff looks ok.
 > 
 > Is it easy to write some sort of automated test?

I couldn't come up with a good way to test this unfortunately, I don't know of any reasonable way to query whether a TCP socket is receiving keepalive messages or to simulate a dropped connection. And for the Python test we would have to ensure the device is connected to the same network as the testing computer which is almost never the case for us. Any ideas would be welcome though, I much prefer to have tests whenever possible.

 > 
 > So if the user previously occasionally had a flaky network
 > connection, they could maintain an adb connection since no
 > keepalives were being used, but now with this change they'll be
 > disconnected since it'll be detected via the missing keepalive
 > pings.
 > 
 > I have no idea if this is acceptable or not (or merely theoretical
 > or too corner-case or we don't care). I'm simply reminded of it
 > because 'back in my day' my friend would keep a ytalk session open
 > with me that would survive disconnecting & reconnecting his dial-up
 > session. Presumably ytalk was not using keepalives back then.


You are correct this would be detrimental for flaky networks; my thinking was that most (all?) adb TCP will be on a local network which should be solid enough to get 1/10 packets. I could add an `adb connect --no-keepalive` option though if we do want to support low-reliability networks. enh@ what do you think?

Patch-set: 1