Update patch set 1
Patch Set 1:

> > Windows stuff looks ok.
 > >
 > > Is it easy to write some sort of automated test?
 > 
 > I couldn't come up with a good way to test this unfortunately, I
 > don't know of any reasonable way to query whether a TCP socket is
 > receiving keepalive messages or to simulate a dropped connection.

Yeah, I see what you mean. I was going to suggest connecting to a socket of another process and kill -9'ing it, but I already know that this won't work on Windows because if a process is hard-killed, the socket is closed and a TCP RST is sent to the other side.

 > You are correct this would be detrimental for flaky networks; my
 > thinking was that most (all?) adb TCP will be on a local network
 > which should be solid enough to get 1/10 packets. I could add an
 > `adb connect --no-keepalive` option though if we do want to support
 > low-reliability networks. enh@ what do you think?

I think another reason I was thinking about this was because bugs like https://code.google.com/p/android/issues/detail?id=191355 make it sound like adb TCP connections can be flaky (though admittedly I don't know the root cause of that specific bug).

Patch-set: 1