Update patch set 2
Patch Set 2:

It would have worked for the trivial case, but it actually calls a different print. Without it the parentheses are simply dropped and the old print builtin is called. With, it will actually call the python 3 compatible function.

Not a problem today, but could easily introduce surprises later.

Patch-set: 2