Update patch set 2
Patch Set 2:

> That's right, but I don't think it's realistic to build bionic
 > against non-Linux kernels.

Android doesn't have non-Linux kernels though, so that doesn't matter to us.

 > > you know no one is maintaining mincrypt, right? that's why i want
 > > to kill it. BoringSSL is what's maintained.
 > 
 > I cc'ed you on the internal thread where Marius confirmed that the
 > Google-internal copy is considered the canonical one that is
 > maintained.

again, that doesn't help Android.

 > Furthermore, external/ projects have README.version
 > files for the purpose of tracking upstream security changes. We
 > have Google-internal tooling that actually parses Android's
 > README.version files and alerts people when a security
 > vulnerability gets discovered in one of the libraries. I hope this
 > addresses your maintenance concerns?

no, because we're the ones who get left holding the baby and doing the work.

(vomit has so far been useless, too, only "alerting" us to a couple of things, and then months after we'd already fixed them. and even if it functioned, it doesn't do the actual work of upgrading.)

Patch-set: 2