Update patch set 3
Patch Set 3: Code-Review-1

We can _assume_ and continue to support the AID_* references from android_filesystem_config.h. These do not need to be deprecated.

The change to getpwnam/getgrnam is to be used when human-readable names are referenced, the relationship between the system AID_* references is defined in android_filesystem_config.h and retains readability.

NB: init may not gain all the benefits of the future adjustments to getpwnam/getgrnam on non-A/B systems. it is run _without_ access to /system/etc/* in the initrd, it has not even been mounted yet. It reads and extends the rules once /system/etc/init/*.rc is mounted and accessible though for partner additions.

Patch-set: 3
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1