Update patch set 1
Patch Set 1:

misplaced reply comment, moved to here:

Tried *NIX [-b|--buffer], '|', '-or-' and 'or', the presentation with 'or' had the most beauty and the least wrap issues.

The idea was to clarify and snug some of the already too long logcat help output so as to add the new feature documented in https://android-review.googlesource.com/#/c/215215

There have been complaints about how long the logcat output is, anything to help keep the irritation of its spew down without resorting to a help hierarchy.

Patch-set: 1