Update patch set 4
Patch Set 4: Code-Review+1

(1 comment)

Is the sysdeps directory new? If so, cool.

I think the Windows APIs already do sanity checks like verifying that you release the critsect on the same thread as it was acquired on. They probably also check for double-release, though I'm not 100% sure.

Patch-set: 4
Label: Code-Review=+1