Update patch set 2
Patch Set 2:

The presubmit failed, because in the mingw32 build, it is not safe to first include <stdatomic.h> and then <atomic>.  Our generic stdatomic.h introduces a bunch of defines that interfere with gcc's <atomic>. For other builds we work around this by having <stdatomic> delegate to <atomic> instead in a C++ environment.  But that doesn't work here because USING_LIBCXX is not defined.

The other include order apparently works.

The conflict currently is with RefBase.h, which uses android_atomic, which includes <stdatomic.h>. go/aog/229218 changes that to <atomic>.

I propose to defer this problem and revisit after submitting 229218 and rebasing.

Patch-set: 2
Label: Verified=0