Update patch set 3
Patch Set 3:

Still testing, but it looks like the mingw build failures indeed mostly disappeared as a result a result of the RefBase check-in.  Very ugly, but I'll take it for now.

There was one more mingw build failure, since the LP64 test for the static_assert was not quite correct.  Windows is LLP64.  Fixed that.

This follows Elliott's advice and does not change alignment. It only adds a static sanity test that clarifies what we expect in terms of alignment.

Reviews welcome.

Patch-set: 3