Update patch set 1
Patch Set 1:

> Cool, sometimes you have to RTFM, sometimes you need to read the
 > code ... will respin

yeah, sorry about that. it's a known bug with the help generation. i've pinged upstream because i think the fact that the help is generally so good makes this an even worse bug!

a good rule of thumb is that anything that deals with ps-like information accepts the same arguments, even if you can't see that in the help text :-(

Patch-set: 1