Update patch set 2
Patch Set 2:

> I am 'nervous' about switching libcutils to cpp, only because one
 > slip and you have a libc++ dependency ... but that said, no
 > problem.
 
you're too late to be worrying about that. try ls *.cpp :-)

and a libc++ dependency would be a *good* thing. far better than more broken/leaky home-made data structures that are too widely used for us to easily remove but require constant security fixes because they're crap. (though libutils is far worse than libcutils.)

 > Such a move to cpp should be done over the entire library in one
 > big rewrite step.

no, that's a mistake. it's always too expensive to do that so it gets left and never happens. mending one window is better than mending none of the windows, even if you're not going to mend them all today.

 > Consider this spaces-fix first, a few bugfixes
 > are pending.

...especially if you're not just rearranging deckchairs and actually have some fixes to do. you'd be surprised how many of the bugs in our old cruft disappear through liberal application of std::string and std::unique_ptr.

 > Do we want to move to 2 spaces indent in this step though?

as you can imagine, that's actually the most contentious issue :-) generally it tends to be that new code gets 2-space indents, but old code stays at 4. but if you're going to touch every line *anyway* because of tabs, there's no reason not to go to 2-space. the history is already muddied.

Patch-set: 2