Update patch set 2
Patch Set 2:

I would agree that the `exec` is heavy weight, but I don't think this is something that we expect to happen often, so I wouldn't be too concerned.

I think we should talk with enh@ and others if we want to change the meaning of on property:a=*.

The last suggestion of having =* cover cases that are not already covered is a big divergence from the current set up where each entry is standalone and I'd recommend against it.  

Maybe we could add a new syntax for =* that specifies "any non-empty" if we don't want to change the existing semantics.

Patch-set: 2