Update patch set 7
Patch Set 7:

Colin managed to sneak in and change the build infrastructure between PS6 and PS7.  The real change is a 1 line addition of RefBase_test to Android.bp instead of Android.mk, and a tiny change to README.txt.  Due to the Android.bp change, this could use another +2.

Patch-set: 7