Update patch set 7
Patch Set 7: Code-Review+2

you're not wrong about "this should be two patches", but we've all got other bugs to deal with... like getting me one of these fancy new kernels :-)

(btw, your "eventually" in your checkin comment is misleading in that this conflict already happened. there are at least four values out there in the wild: 41, 43, 47, and 127. cutils assumes 41, but has no real reason to believe that. the sooner we move to timerslack_ns the better!)

Patch-set: 7
Label: Code-Review=+2