Update patch set 1
Patch Set 1:

Test was in goog/master. No matter, the syntax I was using was beyond grep 101 course level and was likely less maintainable, I prefer KISS when it comes to maintenance, and grep 501 level when I need to make the dog do "back-flips and operate a light switch" :-)

I want to deprecate logpersist script 'trio' in time since it has grown beyond its humble beginning. We now have the Developer Option to control it, and the content is part of the bugreport. However, some test infrastructure is using the command to automate collection; but they can replace all they need it to do with a simple setprop command ...

Patch-set: 1