Update patch set 2
Patch Set 2:

(1 comment)

Will assume that the [\(IF_\)*[A-Z]LOG[A-Z] defines are all OK to add to the environment (which they might not be) will work on refining/testing what we have here.

What do to think about adding a check for ANDROID_NATIVE_API_LEVEL, and for values less than 22 will drop the LOG macros? I could probably also use this level to drop certain interfaces as they do not exist in older liblog.so APIs?

Patch-set: 2
Label: Verified=0