Update patch set 1
Patch Set 1:

Maybe we should have the loopback connect loop through getaddrinfo("localhost"), but have the server try ipv4 loopback first and fallback to ipv6 loopback? That way, we don't break people with hard-coded "127.0.0.1"s, but ipv6 only environments work as expected.


 > no, let's do this right.
 > 
 > doesn't jmgao@ already have a change up to loop through the
 > getaddrinfo results anyway?

That was for generic network connect, not the loopback functions.

Patch-set: 1