Update patch set 5
Patch Set 5: Presubmit-Ready+1 Code-Review+1

Ok, cleaner. The default set of features is everything, so external and partners may wish to place the feature flags, or API version, into their compile options if there is an incompatibility.

Do we have a define in our development environments that actually is set and represents the supported API? AFAIK nothing is set in the compile environments?

I am not 100% comfortable setting api23 and higher for exporting the macros since they have been there forever. Need a straw poll as to this api threshold wisdom.

Patch-set: 5
Label: Presubmit-Ready=+1