Update patch set 2
Patch Set 2:

(3 comments)

It's not that I hate this, but their should only be 1 way to do capabilities on services, either this or using fs capabilities.

If we go this way, fix up dragon to use this method and kill off fs-caps.

The only thing, if this gets merged, i'll be slightly annoyed as I wanted to do this in the beginning versus fs-config-2, but got negative feedback and went that route instead. Granted, fs-config-2 deals with a broader set of issues.

Patch-set: 2