Update patch set 8
Patch Set 8:

> > 0 - I don't think we should merge this without looking at the
 > > alternative approach with fscaps. This is a heavy handed,
 > > non-granular approach to the problem that IMHO, should only be
 > > taken if we know that the exec chains get long and fscaps is too
 > > unwieldy. Jorge gave me an example offline, where the exec chain
 > > was init --> service --> toolbox, so fs caps could have been
 > used,
 > > assuming nothing is on the rootfs (looking at current AOSP I see
 > > nothing).
 > 
 > s/toolbox/helper/ its not in toybox/toolbox.

As I've explained before, no fs cap solution is a drop-in replacement for AID_NET_ADMIN and AID_NET_RAW. Moreover, no fs cap solution allows you to keep your capabilities definition in the .rc file, which allows you to define these things in the same repo as where your process lives.

In light of these facts, and the fact that there's no way to set up file caps from an .rc file, and the fact that we all want to get rid of these patches, but this CL is the first effort to actually get it done, I'd like to move forward with this change.

Patch-set: 8