Update patch set 2
Patch Set 2:

(1 comment)

(The royal) we are working on moving as much as possible from init.rc into individual rc files located at /system/etc/init/ directory. As such, the only content that should reside in init.rc will be about bringup. There is some /proc, /sys and /dev preparations that can not be helped by this. Also all new products should be designed A/B, which means that /system is mounted immediately and then init is run. There is not much need for us to add late bindings as a result.

Patch-set: 2
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>