Update patch set 1
Patch Set 1:

I think your way of testing is interesting and meaningful. I only used small adb commands to test before, and the offline problem was not so obvious. I guess it may be the kernel of host or device keeps the unread data given by adb push, but the data is not what is expected by new adb connection. I raised a bug internally, and will find out if this is the proper solution.

Patch-set: 1
Reviewer: Yabin Cui <1056364@85c56323-6fa9-3386-8a01-6480fb634889>