Update patch set 1
Patch Set 1:

No one uses it because it was inaccessible before, and everyone implemented their own.

When we add the event log tag service (see bugz), we open the door for at least exposing the ability to create binary event logs in the (vapourware) Vendor NDK. More may come. We already have 3 spots in the tree that use this C++ interface for creating binary event logs; more may come.

Patch-set: 1