Update patch set 1
Patch Set 1: Code-Review-1

Well, if I run this on host, I get 6 tests pass, 43 tests fail and tons of noise. I had viewed this as an acceptable response and punshiment to failing to put the --gtest_list_tests flag to the executable. Or do we want to set a good example (?) with a nice clean output of repeating meme of "This test does nothing." and 49 passes? (I do assume the answer is yes and will respin tomorrow).

Patch-set: 1
Label: Code-Review=-1