Update patch set 1
Patch Set 1: Code-Review-2

(3 comments)

I decided on the HAS-A approach because A) this is our C++ code standard, B) we only add interfaces we need and use in the previous CL and C) the code in the previous CL is easier to expand upon and in essence plans for the future.

However, I understand if you feel it is better to write _great_ code at each juncture, and deal with the expansion issues as (or if) they happen.

This code is currently only for demonstration purposes and is merely a prototype that requires other infrastructure (like the logd tag event service) to be useful.

Patch-set: 1
Label: Code-Review=-2