Update patch set 4
Patch Set 4:

This is more of an if you build it, they may come.

enh@ has disagreed with providing a means for applications to proactively ratelimit their messages, as it just means they can hide bad programming practices and bugs. timmurray@ and bdc@ have stated something like the cycles burned are a Q/A signal to take action on bad code. I contend that this is meant to be used inside of library functions that could be called aggressively by _bad_ higher level code, and this is to mitigate logspam and logger DOS (>600 messages/tick). Probably added after the fact as we spot the situations. I expect that whatever causes an aggressive call will remain taking CPU cycles in an even tighter loop and there will remain a strong signal. Abuse of this function hides nothing, but can save EBUSY on the logger sockets, and wasted background cgroup cycles starving logd and others.

Or this will turn into a small piece of unused inert code ...

Patch-set: 4
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>