Update patch set 3
Patch Set 3:

(1 comment)

Subtle: State machine is borken for case where we have three identical message, the intent is to print all three of them as-is, but the result was we stuttered the first message twice, and dropped the third.

Fix in test, and CL forthcoming.

Patch-set: 3
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0