Update patch set 2
Patch Set 2:

> Abandoned
 > 
 > decided to do it in device specific init rc. only device with file
 > based encryption can do this safely.

Hi, Keun-young

Could you please help to describe details of the problem you met here so that we could understand the problem as well?

I enabled this on nougat device without encryption enabled, not see any problem too.(Or I did not find the problem because I don't know what's the problem is).

And in this change, I guess the problem is caused by the moving of zygote. if we only keep the lines that removing settings of cpuset, do you think will it cause any problem on device with DBE?

I think the concept that doing cpuset settings after boot complete is correct, so this change still has something to be merged.


Thanks,
Yongqin Liu

Patch-set: 2